+++ src/include/vitalnix/libvxpdb/libvxpdb.h
@@ -88,7 +88,7 @@ struct vxpdb_driver {
 	const char *name, *desc;
 
 	int (*init)(struct vxpdb_state *, const char *);
-	int (*open)(struct vxpdb_state *, long);
+	int (*open)(struct vxpdb_state *, unsigned int);
 	void (*close)(struct vxpdb_state *);
 	void (*exit)(struct vxpdb_state *);
